+++ database/migrations/2024_02_12_212530_create_dpsm_akontacijes_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('sifra_vrste_placanja')->nullable();
             $table->string('SLOV_grupa_vrste_placanja')->nullable();
             $table->integer('sati')->nullable();
-            $table->integer('iznos')->nullable();
+            $table->float('iznos',15,4)->nullable();
             $table->integer('procenat')->nullable();
             $table->string('POK2_obracun_minulog_rada')->nullable();
